+++ Manager/JobManagerInterface.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Dtc\QueueBundle\Model;
+namespace Dtc\QueueBundle\Manager;
 
 interface JobManagerInterface
 {
-    public function resetErroneousJobs($workerName = null, $methodName = null);
+    public function resetExceptionJobs($workerName = null, $methodName = null);
 
-    public function pruneErroneousJobs($workerName = null, $methodName = null);
+    public function pruneExceptionJobs($workerName = null, $methodName = null);
 
     /**
      * Prunes (or archived) jobs that are expired.
@@ -25,6 +25,11 @@ interface JobManagerInterface
 
     public function save(Job $job);
 
+    /**
+     * Called after a job has finished - may delete the job / reset the job and/or do other related cleanup
+     * @param Job $job
+     * @return bool Whether the job is deletable or not (true - job can be deleted, false - job was reset or reused in some way)
+     */
     public function saveHistory(Job $job);
 
     public function resetStalledJobs($workerName = null, $method = null);